assistant/src/requirements.txt
pytest==6.1.2
pytest-cov==3.0.0
Flask==2.0.2
flake8==4.0.1
mypy==0.931
pylint==2.12.2
isort==5.10.1
safety==1.10.3
gunicorn==20.1.0
requests==2.27.1
google-assistant-grpc==0.2.1
google-auth-oauthlib>=0.1.0,<0.3
google-cloud-speech==2.13.1
urllib3[secure]>=1.21,<2
sounddevice>=0.3.7,<0.4
tenacity==5.0.4
pathlib2>=2.3.0,<3
click>=7.1.2
